+++ apiphany-core/src/main/java/org/apiphany/client/http/AbstractHttpExchangeClient.java
@@ -20,8 +20,8 @@ import org.apiphany.http.HttpHeader;
 import org.apiphany.http.HttpHeaderValues;
 import org.apiphany.http.HttpStatus;
 import org.apiphany.io.ContentType;
-import org.apiphany.json.JsonBuilder;
 import org.apiphany.json.jackson2.Jackson2JsonHttpContentConverter;
+import org.apiphany.json.jackson2.Jackson2Library;
 import org.apiphany.security.ssl.SSLContexts;
 import org.apiphany.security.ssl.SSLProperties;
 import org.morphix.lang.JavaObjects;
@@ -92,7 +92,7 @@ public abstract class AbstractHttpExchangeClient implements HttpExchangeClient {
 	public static void addDefaultContentConverters(final List<ContentConverter<?>> contentConverters) {
 		contentConverters.add(new StringHttpContentConverter());
 
-		if (JsonBuilder.isJacksonPresent()) {
+		if (Jackson2Library.isPresent()) {
 			contentConverters.add(new Jackson2JsonHttpContentConverter<>());
 		}
 	}